+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  - New Job Engine Lite Docker image build, optimized for running jobs from within an infrastructure, with multi-arch support
  - When a job with execution-mode=mixed fails, it get back in the job queue in pull-mode
  - Create deployment_state jobs with execution-mode 
+ - Add pause entrypoint
  
 ### Changed